remove support for non-sql migrations
Tying non-SQL migrations to a migration tool has several bad
side-effects, but the biggest is that dependencies of the underlying
scripts may change requiring changes to the scripts themselves, so the
hash of the script doesn't have value/meaning. Critically, scripts that
ran for one developer yesterday might break for a different developer
today, so it's impossible to achieve consistent state with this
approach if using dependencies (and almost any script accessing the
database would).